Api-Netural/requirements.txt
# requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
torch==2.3.0
numpy==1.26.0
pydantic==2.7.0
pydantic-settings==2.3.0
sqlalchemy==2.0.30
pyodbc==5.1.0
pymssql==2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
alembic==1.13.0
asyncio-mqtt==0.16.0
pandas==2.2.3
scikit-learn==1.6.1
openpyxl==3.1.5
python-multipart==0.0.20
websockets==14.1
joblib==1.4.2